+++ lib/main.dart
@@ -12,7 +12,8 @@ class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return GetMaterialApp(
+    return MaterialApp(
+        navigatorKey: Get.key,
         title: 'Flutter Demo',
         theme: ThemeData(
           primarySwatch: Colors.blue,