+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG.md
 
+## 1.0.1 (2021-11-4)
+
+Changes:
+
+- Payment now fully supports test keys from www.coinforbarter.customize
+- Customizable button button color and text color
+
 ## 1.0.0 (2021-11-3)
 
 Features: